services/bamul-app-api/pyproject.toml
[project]
name = "bamul-app-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "arq>=0.27.0",
    "asyncpg>=0.31.0",
    "boto3>=1.42.89",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.13.1",
    "python-jose[cryptography]>=3.5.0",
    "redis[asyncio]>=5.3.1",
    "sqlalchemy[asyncio]>=2.0.49",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.44.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.10",
]

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["app", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "S", "B", "A", "C4", "PT", "RET", "SIM", "PLW", "PLC"]
ignore = ["S101"]  # assert allowed in tests

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S", "N802"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
